% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_data}
\alias{dummy_data}
\title{Dummy dataset}
\format{A dataframe with 1,039 rows and 5 variables
\describe{
  \item{STRATUM}{sampling strata}
  \item{TREATMENT}{participant names}
  \item{RESPONSE_1}{numeric response variable 1}
  \item{RESPONSE_2}{numeric response variable 2}
  \item{DATE}{date of measurements with YYYYMMDD format}
}}
\usage{
dummy_data
}
\description{
A dummy dataset
}
\keyword{datasets}
